# CatalogoParametrosUcoLab/docs/openapi/catalogo-parametros-v1.yaml
openapi: 3.0.3
info:
  title: Catálogo de Parámetros UcoLab
  version: 1.0.0
  description: |
    Contrato público y versionado del módulo Catálogo de Parámetros.
    Este archivo es la fuente de verdad para productores y consumidores.
servers:
  - url: http://localhost:8080
    description: Desarrollo local
...
paths:
...
  /catalogo-parametros/api/v1/modulos:
    get:
      tags: [Módulos]
      operationId: consultarModulos
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { $ref: '#/components/responses/ModulosOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [Módulos]
      operationId: crearModulo
      requestBody: { $ref: '#/components/requestBodies/ModuloRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /catalogo-parametros/api/v1/modulos/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Módulos]
      operationId: consultarModuloPorId
      responses:
        '200': { $ref: '#/components/responses/ModulosOk' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [Módulos]
      operationId: actualizarModulo
      requestBody: { $ref: '#/components/requestBodies/ModuloRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Módulos]
      operationId: eliminarModulo
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /catalogo-parametros/api/v1/modulos/{id}/changestatus:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Módulos]
      operationId: cambiarEstadoModulo
      summary: Cambia únicamente el estado activo del módulo
      requestBody: { $ref: '#/components/requestBodies/CambiarEstadoRequest' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador UUID del recurso
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      description: Página basada en uno
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Cantidad máxima de registros solicitados
      schema: { type: integer, minimum: 1, default: 10 }
  requestBodies:
    CambiarEstadoRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CambiarEstadoRequest' }
...
    ModuloRequest:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ModuloRequest' }
...
  responses:
    Success:
      description: Operación exitosa
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    Created:
      description: Recurso creado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    BadRequest:
      description: Petición inválida o regla de negocio incumplida
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    NotFound:
      description: Recurso no encontrado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    Conflict:
      description: Conflicto con el estado actual del catálogo
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
    InternalError:
      description: Error técnico no esperado
      content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } }
...
    ModulosOk: { description: Módulos encontrados, content: { application/json: { schema: { $ref: '#/components/schemas/ModuloResponse' } } } }
...
  schemas:
    CambiarEstadoRequest:
      type: object
      required: [activo]
      properties:
        activo: { type: boolean, example: false }
    Uuid:
      type: string
      format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
    InputUuid:
      type: string
      format: uuid
      description: Los DTO actuales reciben el UUID como cadena JSON.
      example: 123e4567-e89b-12d3-a456-426614174000
    InputBoolean:
      type: string
      enum: ['true', 'false']
      default: 'true'
      description: Los DTO actuales reciben el booleano como cadena JSON.
    InputDateTime:
      type: string
      format: date-time
      description: Fecha y hora ISO-8601 con desplazamiento obligatorio; se recomienda enviar la zona de Colombia como -05:00.
      example: '2026-09-11T08:30:00-05:00'
    MessageResponse:
      type: object
      required: [mensajes]
      properties:
        mensajes: { type: array, items: { type: string } }
...
    ModuloRequest:
      type: object
      additionalProperties: false
      required: [nombre, idAplicacion, activo, fechaInicio, fechaFinal]
      properties:
        nombre: { type: string, minLength: 1, example: Matrículas }
        idAplicacion: { $ref: '#/components/schemas/InputUuid' }
        activo: { $ref: '#/components/schemas/InputBoolean' }
        fechaInicio: { $ref: '#/components/schemas/InputDateTime' }
        fechaFinal: { $ref: '#/components/schemas/InputDateTime' }
...
    Modulo:
      type: object
      required: [id, nombre, idAplicacion, activo]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        nombre: { type: string }
        idAplicacion: { $ref: '#/components/schemas/Uuid' }
        activo: { type: boolean }
        fechaInicio: { type: string, format: date-time, nullable: true, example: '2026-09-11T13:30:00Z' }
        fechaFinal: { type: string, format: date-time, nullable: true, example: '2026-12-31T04:59:59Z' }
...
    ModuloResponse: { allOf: [{ $ref: '#/components/schemas/MessageResponse' }, { type: object, required: [modulos], properties: { modulos: { type: array, items: { $ref: '#/components/schemas/Modulo' } } } }] }
...